Source/Minor/Fix: Fix for invalid generated favicon
There can be a difference in the extension between
the predefined favicon and the provided favicon.
This difference has to be accounted for.

Assets imports have to be prefixed with an initial
forward slash before they can be resolved to the
public directory.
E.g asset/image/favicon.png has to be written
as: /asset/image/favicon.png

Fix for multiple occurrence of '>' before a '<' is
found in escape JSX. We have to be as close as
possible to avoid being in a quotation context.