Source/Minor: Extend page name to include underscore and numbers
Page names can be further disambiguated with numbers
and underscore

Assets already present in /assets folder are placed
in a single /assets folder instead of /assets/assets
which was previously used.

Addition of defer attribute to all scripts so as to
delay their load time and synchronize them with
their point of usage.

Multiline quotes that have not been previously
escaped are now being escaped in JSX style.