Source(feat): Read project from archive
Archives can now be specified to the project.
This will be extracted and placed in the
os temporary directory. The extracted archive
will be deleted once the generation is complete.

The presence/absence of an archive's root
directory is detected and also resolved.

A new option entryPoint has been provided to
select the starting index file. It defaults
to index.html.