Source/fix: Update isURI to account for windows path
isURI tries to construct a `new URI` and makes decision
based on if an exception is thrown. This underlying
function treats windows path as a valid url. A separate
check is used to check if the path is emit properties
of a windows path (such as beginning with a drive
letter).