Source/Minor: Add page reload on empty anchor tags
When an anchor tag having an href property of the
form href="" is clicked, the page reloads. We
need to prevent default on this type of anchors.
We implement the onClick function for the anchor
tag to achieve this.

In the findIndexFile function, the code iterating
through the directoryQueue ought to check the
returned file from the impl function which should
be returned appropriately if it is defined.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>