Source/Fix: Fix bug in comment generation code
When converting html style comments to js,
the content of <!--...--> are picked and
converted into {/*...*/}. If the content
starts with a /, it will mess up our
generation. Hence, we pad our comment in
this form {/* ... */}.

A bug creeped into expandMatches function
where an early return was used instead of
a continue statement.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>