Source/Fix: Code annotation and bug fixes
To prevent download path from redirecting
infinitely, we implement a depth counter.
Once the counter is exceeded, we enter
an error state.

Fix for isSuperSetOf() method. The params.
have to be sorted. We use two cursors, one
for the supposed superset and the other
for the reference. We move the superset
always but we only move the reference on
equality with the current value held at
the current cursor of the superset.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>