Source/Minor: Rewrite rule for asset resolution
Assets are now resolved only by their rela property.
Hence if the html file has an incorrect asset
link, we won't be able to resolve it. Had we resolve
it, we could have selected the wrong asset as we
would select the closest match to that current
processing directory.

Inplace urls in styles (e.g background-image: url(...))
are now resolved directly by the formatStyle function.

We test if a file is binary before trying to search
for assets in it.

The getRootDirectory function uses the user provided
path in addition to the path where the entryPoint is
found to calculate the root directory.
For example:
User provided path to search for entryPoint is
/home/user/html-template/ -> initialPath
We found the entryPoint file in
/home/user/html-template/source/pages/ -> mainDir
For here, it is evident that we have to select the
`initialPath` as the root path as it is the parent of the
`mainDir` path

A new flag `weakReplacement` is provided to guide projects
that depends on the links in their `href` attribute in
the source code. Turing on this flag prevents the
`href`'s content from being replace with
`javascript:void(0);'.

The session temporary directory
`${system temporary dir}/ReactifyHTML/${sessionID}` is now
cleaned up properly.

The useScript hook path resolution has now being fixed.

HTML comments containing Js commented texts now have their
Js comments escaped before being converted to JSX style
comment.
Example:
<!-- /* This is the landing page */ -->
Is converted to
{/* /* This is the landing page *\/ */}

Implicit references (e.g /docs/api -> /docs/api/index.html)
have been extended to detect query and fragment parameters.
Example:
link: /docs/api/?v1.02#home
Has base as /docs/api/, query param as ?v1.02, and fragment,
\#home. The base is used in deciding if a path is an
implicit reference.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>