Source/feat: Implementation of new script loader
We now load scripts using a component-loader in
each page. The component-loader mounts and unmounts
the scripts on page navigation. Page metas and
links are loaded this way also. We want our
generated react script to be as close to the
html source as possible so as not to introduce
errors such as the js not being able to pick out
an HTMLElement as it normally would in a pure
HTML project.

All scripts are wrapped in a IIFE. We don't
want variables to escape the scope in which
they are declared.

A new option has been added (weakReplacement).
Some scripts rely on the href attributes of
anchor tags. If you have such a project,
turning on this flag would not replace the
href attribute with `javascript:void(0)`.

The hooks mechanism has been replaced with
the component-loader mechanism.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>