Source/Fix: Fix for crashing view from navigator
A custom hook `useNavigator` has been designed
in order to refactor the use of `navigateTo`
in earlier commits. This patch also include
fix for the `fixAnchorRoutes` function which
erroneously used the current page name when
resolving anchor page links making the
program stay on the same page. We transformed
the routeMap value into a pair setting the
second property to the page associated with
the link. We use the information from this
page in `fixAnchorRoutes` to derive the
page name associated with the currently
processed route.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>